move TooltipIcon components out of Geral subfolder
